pyproject.toml
[project]
name = "babykicks"
version = "0.1.0"
description = "Baby kick counter"
authors = [
    {name = "Syafiq Termizi",email = "ahmadsyafiq93@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.13,<0.116.0)",
    "pydantic-settings (>=2.10.0,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.development.dependencies]
ruff = "^0.12.0"
black = "^25.1.0"
